feat(game): add GameHud component and EditModeGrid indicator
Introduce a new GameHud component to consolidate and organize all HUD
elements in the game UI, improving structure and maintainability.

Add EditModeGrid component to display a helper during edit mode,
with colors adapting to time of day for better visual context.

Update dynamic import in GameSceneDynamic to correctly load
GameSceneWrapper component.

Adjust ShoppingCartHud toggle button focus ring offset for improved
accessibility styling.

Add tailwindcss-animate plugin to Tailwind config to enable animation
utilities.
