+++ apps/www/package.json
@@ -57,6 +57,7 @@
     "postcss": "8.5.6",
     "react-markdown": "10.1.0",
     "tailwindcss": "3.4.17",
+    "tailwindcss-animate": "1.0.7",
     "typescript": "5.8.3"
   }
 }
