+++ packages/game/src/GameSceneDynamic.tsx
@@ -3,7 +3,7 @@
 import { GardenLoadingIndicator } from './indicators/GardenLoadingIndicator';
 import dynamic from 'next/dynamic';
 
-export const GameSceneDynamic = dynamic(() => import('./GameScene').then(mod => mod.GameSceneWrapper), {
+export const GameSceneDynamic = dynamic(() => import('./GameSceneWrapper').then(mod => mod.GameSceneWrapper), {
     ssr: false,
     loading: () => <GardenLoadingIndicator />
 });
