+++ packages/game/src/indicators/GardenLoadingIndicator.tsx
@@ -2,7 +2,7 @@ import { cx } from '@signalco/ui-primitives/cx';
 import styles from './GardenLoadingIndicator.module.scss';
 import { HTMLAttributes } from 'react';
 
-export function GardenLoadingIndicator({className, ...rest}: HTMLAttributes<HTMLDivElement>) {
+export function GardenLoadingIndicator({ className, ...rest }: HTMLAttributes<HTMLDivElement>) {
     return (
         <div className={cx('absolute top-1/2 left-1/2 -translate-x-1/2', className)} {...rest}>
             <div className={styles.boxes}>
